Add type to enumerate all positions in grid
The type is called `AllPositionsInGrid`.
This kind of enumeration is frequently done in a pattern of two nested `for`s,
so use of this type will cut on duplication of that pattern.

The iterable returned by the type is initialized with an invalid position row: -1, column: grid size - 1,
so that it is not necessary to handle this inital state specially in `next()`.